Move Voigt1D from sherpa.models.basic to sherpa.astro.models
I took the opportunity to rename sherpa/astro/models/tests/test_models.py
to .../test_astro_models.py as it better matches our naming scheme.

I also changed the regression test for Voigt1D to pick parameter
valyes that better match the x-axis range (ie fwhm from 10 to 0.3).
