+++ docs/model_classes/astro_models.rst
@@ -7,10 +7,10 @@ The sherpa.astro.models module
 .. automodule:: sherpa.astro.models
 
    .. rubric:: Classes
-               
+
    .. autosummary::
       :toctree: api
-   
+
       Atten
       BBody
       BBodyFreq
@@ -31,9 +31,10 @@ The sherpa.astro.models module
       Schechter
       Sersic2D
       Shell2D
+      Voigt1D
 
 Class Inheritance Diagram
 =========================
 
-.. inheritance-diagram:: Atten BBody BBodyFreq BPL1D Beta1D Beta2D DeVaucouleurs2D Dered Disk2D Edge HubbleReynolds JDPileup LineBroad Lorentz1D Lorentz2D MultiResponseSumModel NormBeta1D Schechter Sersic2D Shell2D
+.. inheritance-diagram:: Atten BBody BBodyFreq BPL1D Beta1D Beta2D DeVaucouleurs2D Dered Disk2D Edge HubbleReynolds JDPileup LineBroad Lorentz1D Lorentz2D MultiResponseSumModel NormBeta1D Schechter Sersic2D Shell2D Voigt1D
    :parts: 1
