+++ docs/model_classes/basic.rst
@@ -36,7 +36,6 @@ The sherpa.models.basic module
       TableModel
       Tan
       UserModel
-      Voigt1D
 
       Box2D
       Const2D
@@ -50,5 +49,5 @@ The sherpa.models.basic module
 Class Inheritance Diagram
 =========================
 
-.. inheritance-diagram:: Box1D Const1D Scale1D Cos Delta1D Erf Erfc Exp Exp10 Gauss1D Integrate1D Log Log10 LogParabola NormGauss1D Poisson Polynom1D PowLaw1D Sin Sqrt StepHi1D StepLo1D TableModel Tan UserModel Voigt1D Box2D Const2D Scale2D Delta2D Gauss2D SigmaGauss2D NormGauss2D Polynom2D
+.. inheritance-diagram:: Box1D Const1D Scale1D Cos Delta1D Erf Erfc Exp Exp10 Gauss1D Integrate1D Log Log10 LogParabola NormGauss1D Poisson Polynom1D PowLaw1D Sin Sqrt StepHi1D StepLo1D TableModel Tan UserModel Box2D Const2D Scale2D Delta2D Gauss2D SigmaGauss2D NormGauss2D Polynom2D
    :parts: 1
