+++ helpers/extensions/__init__.py
@@ -19,9 +19,6 @@
 #
 
 
-import shlex
-import os
-
 from numpy.distutils.core import Extension
 
 # Include directory for Sherpa headers
@@ -48,10 +45,10 @@ def get_deps(deps):
     alldeps = set()
 
     while deps:
-        next = deps.pop()
-        if next not in alldeps:
-            alldeps.add(next)
-            deps.update(header_deps[next])
+        dep = deps.pop()
+        if dep not in alldeps:
+            alldeps.add(dep)
+            deps.update(header_deps[dep])
 
     return [sherpa_inc[0] + '/sherpa/' + d + '.hh' for d in alldeps]
 
@@ -153,11 +150,9 @@ utils = Extension('sherpa.utils._utils',
                         'sherpa/utils/src/sjohnson/Faddeeva.hh']))
 
 modelfcts = Extension('sherpa.models._modelfcts',
-              ['sherpa/models/src/_modelfcts.cc',
-               'sherpa/utils/src/sjohnson/Faddeeva.cc'],
-                      sherpa_inc +['sherpa/utils/src/sjohnson',],
-                      depends=get_deps(['model_extension', 'models'])+
-                      ['sherpa/utils/src/sjohnson/Faddeeva.hh'])
+                      ['sherpa/models/src/_modelfcts.cc'],
+                      sherpa_inc,
+                      depends=get_deps(['model_extension', 'models']))
 
 saoopt = Extension('sherpa.optmethods._saoopt',
               ['sherpa/optmethods/src/_saoopt.cc',
@@ -223,9 +218,11 @@ integration = Extension('sherpa.utils.integration',
                         'sherpa/utils/src/gsl/gsl_integration.h']))
 
 astro_modelfcts = Extension('sherpa.astro.models._modelfcts',
-              ['sherpa/astro/models/src/_modelfcts.cc'],
-              sherpa_inc,
-              depends=get_deps(['model_extension', 'astro/models']))
+                            ['sherpa/astro/models/src/_modelfcts.cc',
+                             'sherpa/utils/src/sjohnson/Faddeeva.cc'],
+                            sherpa_inc + ['sherpa/utils/src/sjohnson'],
+                            depends=get_deps(['model_extension', 'astro/models']) + \
+                                    ['sherpa/utils/src/sjohnson/Faddeeva.hh'])
 
 pileup = Extension('sherpa.astro.utils._pileup',
               ['sherpa/astro/utils/src/fftn.c',
