+++ sherpa/models/src/_modelfcts.cc
@@ -1,4 +1,4 @@
-// 
+//
 //  Copyright (C) 2010, 2020  Smithsonian Astrophysical Observatory
 //
 //
@@ -47,8 +47,7 @@ static PyMethodDef ModelFcts[] = {
   MODELFCT1D( stephi1d, 2 ),
   MODELFCT1D( steplo1d, 2 ),
   MODELFCT1D( tan, 3 ),
-  MODELFCT1D_NOINT( wofz, 4 ),
-  
+
   MODELFCT2D( box2d, 5 ),
   MODELFCT2D( const2d, 1 ),
   MODELFCT2D( delta2d, 3 ),
@@ -56,7 +55,7 @@ static PyMethodDef ModelFcts[] = {
   MODELFCT2D_NOINT( sigmagauss2d, 6 ),
   MODELFCT2D_NOINT( ngauss2d, 6 ),
   MODELFCT2D( poly2d, 9 ),
-  
+
   PY_MODELFCT1D_INT((char*)"integrate1d",
 		 (char*)"integrate user functions\n\nExample:\n int_array = integrate1d(func, param_array, xlo_array, xhi_array)" ),
