+++ sherpa/models/tests/test_basic.py
@@ -19,7 +19,6 @@
 #
 
 import numpy as np
-from pytest import approx
 
 import sherpa.models.basic as basic
 from sherpa.utils import SherpaFloat, _utils
@@ -30,72 +29,49 @@ def userfunc(pars, x, *args, **kwargs):
     return x
 
 
-class TestBasic:
-    excluded_models = (ArithmeticModel, RegriddableModel1D, RegriddableModel2D, basic.Const)
+EXCLUDED_MODELS = (ArithmeticModel, RegriddableModel1D, RegriddableModel2D, basic.Const)
 
-    def test_create_and_evaluate(self):
-        x = np.arange(1.0, 5.0)
-        count = 0
+def test_create_and_evaluate():
 
-        for cls in dir(basic):
-            clsobj = getattr(basic, cls)
+    x = np.arange(1.0, 5.0)
+    count = 0
 
-            if not isinstance(clsobj, type) \
-                or not issubclass(clsobj, ArithmeticModel) \
-                or clsobj in self.excluded_models:
-                continue
+    for cls in dir(basic):
+        clsobj = getattr(basic, cls)
 
-            # These have very different interfaces than the others
-            if cls == 'Integrator1D' or cls == 'Integrate1D':
-                continue
+        if not isinstance(clsobj, type) \
+            or not issubclass(clsobj, ArithmeticModel) \
+            or clsobj in EXCLUDED_MODELS:
+            continue
 
-            m = clsobj()
-            if isinstance(m, basic.TableModel):
-                m.load(x,x)
-            if isinstance(m, basic.UserModel):
-                m.calc = userfunc
-            assert type(m).__name__.lower() == m.name
-            count += 1
+        # These have very different interfaces than the others
+        if cls == 'Integrator1D' or cls == 'Integrate1D':
+            continue
 
-            try:
-                if m.name.count('2d'):
-                    pt_out  = m(x, x)
-                    int_out = m(x, x, x, x)
+        m = clsobj()
+        if isinstance(m, basic.TableModel):
+            m.load(x,x)
+        if isinstance(m, basic.UserModel):
+            m.calc = userfunc
+        assert type(m).__name__.lower() == m.name
+        count += 1
+
+        try:
+            if m.name.count('2d'):
+                pt_out  = m(x, x)
+                int_out = m(x, x, x, x)
+            else:
+                if m.name in ('log', 'log10'):
+                    xx = -x
                 else:
-                    if m.name in ('log', 'log10'):
-                        xx = -x
-                    else:
-                        xx = x
-                    pt_out  = m(xx)
-                    int_out = m(xx, xx)
-            except ValueError:
-                self.fail("evaluation of model '%s' failed" % cls)
+                    xx = x
+                pt_out  = m(xx)
+                int_out = m(xx, xx)
+        except ValueError:
+            assert False, "evaluation of model '{}' failed".format(cls)
 
-            for out in (pt_out, int_out):
-                assert out.dtype.type is SherpaFloat
-                assert out.shape == x.shape
+        for out in (pt_out, int_out):
+            assert out.dtype.type is SherpaFloat
+            assert out.shape == x.shape
 
-        assert count == 33
-
-
-def test_basic():
-    my_test_basic = TestBasic()
-    my_test_basic.test_create_and_evaluate()
-
-
-class TestVoigt1D:
-
-    def test_voigt(self):
-    
-        x = np.linspace(-0.8, 0.8, 10)
-        voigt = basic.Voigt1D()
-        voigt_result = voigt(x)
-
-        result = [0.04457704, 0.04470863, 0.04480765, 0.04487382, 0.04490695,
-                  0.04490695, 0.04487382, 0.04480765, 0.04470863, 0.04457704]
-
-        assert voigt_result == approx(result)
-
-def test_voigt():
-    my_test_voigt = TestVoigt1D()
-    my_test_voigt.test_voigt()
+    assert count == 32
